OPHAKTKEH-200 Use skeletons and refactor styles
