+++ src/main/reactjs/src/components/clerkTranslator/ClerkTranslatorFilters.tsx
@@ -1,5 +1,5 @@
 import SearchIcon from '@mui/icons-material/Search';
-import { Button, InputAdornment } from '@mui/material';
+import { InputAdornment } from '@mui/material';
 import { useEffect, useState } from 'react';
 
 import { ComboBox, valueAsOption } from 'components/elements/ComboBox';
@@ -14,60 +14,12 @@ import {
   useKoodistoLanguagesTranslation,
 } from 'configs/i18n';
 import { useAppDispatch, useAppSelector } from 'configs/redux';
-import { Color, TextFieldVariant, Variant } from 'enums/app';
-import { AuthorisationStatus } from 'enums/clerkTranslator';
+import { TextFieldVariant } from 'enums/app';
 import { useDebouncedValue } from 'hooks/useDebouncedValue';
 import { ClerkTranslatorFilter } from 'interfaces/clerkTranslator';
 import { AutocompleteValue } from 'interfaces/combobox';
 import { setClerkTranslatorFilters } from 'redux/actions/clerkTranslator';
-import {
-  clerkTranslatorsSelector,
-  selectTranslatorsByAuthorisationStatus,
-} from 'redux/selectors/clerkTranslator';
-
-export const RegisterControls = () => {
-  const { authorised, expiring, expired } = useAppSelector(
-    selectTranslatorsByAuthorisationStatus
-  );
-  const { t } = useAppTranslation({
-    keyPrefix: 'akt.component.clerkTranslatorFilters.authorisationStatus',
-  });
-  const dispatch = useAppDispatch();
-  const { filters } = useAppSelector(clerkTranslatorsSelector);
-  const filterByAuthorisationStatus = (status: AuthorisationStatus) => {
-    dispatch(setClerkTranslatorFilters({ authorisationStatus: status }));
-  };
-  const variantForStatus = (status: AuthorisationStatus) => {
-    return status === filters.authorisationStatus
-      ? Variant.Contained
-      : Variant.Outlined;
-  };
-
-  const countsForStatuses = [
-    { status: AuthorisationStatus.Authorised, count: authorised.length },
-    { status: AuthorisationStatus.Expiring, count: expiring.length },
-    { status: AuthorisationStatus.Expired, count: expired.length },
-  ];
-
-  return (
-    <>
-      {countsForStatuses.map(({ count, status }, i) => (
-        <Button
-          key={i}
-          data-testid={`clerk-translator-filters__btn--${status}`}
-          color={Color.Secondary}
-          variant={variantForStatus(status)}
-          onClick={() => filterByAuthorisationStatus(status)}
-        >
-          <div className="columns gapped">
-            <div className="grow">{t(status)}</div>
-            <div>{`(${count})`}</div>
-          </div>
-        </Button>
-      ))}
-    </>
-  );
-};
+import { clerkTranslatorsSelector } from 'redux/selectors/clerkTranslator';
 
 export const ClerkTranslatorFilters = () => {
   // i18n
@@ -99,64 +51,66 @@ export const ClerkTranslatorFilters = () => {
   }, [debouncedName, dispatch]);
 
   return (
-    <div className="grid-columns gapped">
-      <div className="rows gapped-xxs">
-        <H3>{t('languagePair.title')}</H3>
-        <div className="columns gapped">
-          <LanguageSelect
-            autoHighlight
-            data-testid="clerk-translator-filters__from-lang"
-            label={t('languagePair.fromPlaceholder')}
-            excludedLanguage={filters.toLang}
-            value={getLanguageSelectValue(filters.fromLang)}
-            languages={langs.from}
-            variant={TextFieldVariant.Outlined}
-            onChange={handleFilterChange('fromLang')}
+    <div className="rows gapped">
+      <div className="grid-columns gapped">
+        <div className="rows gapped-xs">
+          <H3>{t('languagePair.title')}</H3>
+          <div className="columns gapped">
+            <LanguageSelect
+              autoHighlight
+              data-testid="clerk-translator-filters__from-lang"
+              label={t('languagePair.fromPlaceholder')}
+              excludedLanguage={filters.toLang}
+              value={getLanguageSelectValue(filters.fromLang)}
+              languages={langs.from}
+              variant={TextFieldVariant.Outlined}
+              onChange={handleFilterChange('fromLang')}
+            />
+            <LanguageSelect
+              autoHighlight
+              data-testid="clerk-translator-filters__to-lang"
+              label={t('languagePair.toPlaceholder')}
+              excludedLanguage={filters.fromLang}
+              value={getLanguageSelectValue(filters.toLang)}
+              languages={langs.to}
+              variant={TextFieldVariant.Outlined}
+              onChange={handleFilterChange('toLang')}
+            />
+          </div>
+        </div>
+        <div className="rows gapped-xs">
+          <H3>{t('name.title')}</H3>
+          <CustomTextField
+            data-testid="clerk-translator-filters__name"
+            label={t('name.placeholder')}
+            type="search"
+            InputProps={{
+              endAdornment: (
+                <InputAdornment position="end">
+                  <SearchIcon />
+                </InputAdornment>
+              ),
+            }}
+            value={name}
+            onChange={(event) => {
+              setName(event.target.value);
+            }}
           />
-          <LanguageSelect
+        </div>
+        <div className="rows gapped-xs">
+          <H3>{t('town.title')}</H3>
+          <ComboBox
             autoHighlight
-            data-testid="clerk-translator-filters__to-lang"
-            label={t('languagePair.toPlaceholder')}
-            excludedLanguage={filters.fromLang}
-            value={getLanguageSelectValue(filters.toLang)}
-            languages={langs.to}
+            data-testid="clerk-translator-filters__town"
+            label={t('town.placeholder')}
+            values={towns.map(valueAsOption)}
+            value={filters.town ? valueAsOption(filters.town) : null}
             variant={TextFieldVariant.Outlined}
-            onChange={handleFilterChange('toLang')}
+            onChange={handleFilterChange('town')}
           />
         </div>
       </div>
-      <div className="rows gapped-xxs">
-        <H3>{t('name.title')}</H3>
-        <CustomTextField
-          data-testid="clerk-translator-filters__name"
-          label={t('name.placeholder')}
-          type="search"
-          InputProps={{
-            endAdornment: (
-              <InputAdornment position="end">
-                <SearchIcon />
-              </InputAdornment>
-            ),
-          }}
-          value={name}
-          onChange={(event) => {
-            setName(event.target.value);
-          }}
-        />
-      </div>
-      <div className="rows gapped-xxs">
-        <H3>{t('town.title')}</H3>
-        <ComboBox
-          autoHighlight
-          data-testid="clerk-translator-filters__town"
-          label={t('town.placeholder')}
-          values={towns.map(valueAsOption)}
-          value={filters.town ? valueAsOption(filters.town) : null}
-          variant={TextFieldVariant.Outlined}
-          onChange={handleFilterChange('town')}
-        />
-      </div>
-      <div className="rows gapped-xxs">
+      <div className="rows gapped-xs">
         <H3>{t('authorisationBasis.title')}</H3>
         <ComboBox
           autoHighlight
