+++ src/main/reactjs/src/components/clerkTranslator/clerkTranslatorDetails/AuthorisationDetails.tsx
@@ -9,7 +9,7 @@ import {
   TableRow,
 } from '@mui/material';
 
-import { H2, Text } from 'components/elements/Text';
+import { H3, Text } from 'components/elements/Text';
 import {
   useAppTranslation,
   useKoodistoLanguagesTranslation,
@@ -32,8 +32,8 @@ export const AuthorisationDetails = ({
 
   return (
     <>
-      <div className="columns">
-        <H2 className="grow">{t('header')}</H2>
+      <div className="columns margin-top-xxl">
+        <H3 className="grow">{t('header')}</H3>
         <Button
           data-testid="clerk-translator-details__authorisation-details__add-btn"
           variant={Variant.Contained}
