+++ src/main/reactjs/src/styles/abstracts/common/_spacing.scss
@@ -1,7 +1,7 @@
 /* Widths */
 .full-max-width {
   &#{&} {
-    max-width: 100% !important;
+    max-width: 100%;
   }
 }
 
@@ -9,9 +9,20 @@
   max-width: 50%;
 }
 
+/* Heights */
+.full-height {
+  &#{&} {
+    height: 100vh;
+  }
+}
+
 /* Margins */
 .margin-top-sm {
   margin-top: map-get($spacing, 'sm');
+}
+
+.margin-top-lg {
+  margin-top: map-get($spacing, 'lg');
 }
 
 .margin-top-xxl {
@@ -25,6 +36,10 @@
 
 .gapped-sm {
   gap: map-get($spacing, 'sm');
+}
+
+.gapped-xs {
+  gap: map-get($spacing, 'xs');
 }
 
 .gapped-xxs {
